Set title and rewrite introduction.
We need to start with a clear explanation of *why* we're doing
everything we'll be presenting in the blog post. And the message here
needs to be simple and clear: we're just saying that we want to track
benchmark results over time. Generalities about benchmarking being
hard aren't really the point, because we're not really making any
existing workflow easier, just enabling new ones.