Som small edits.
Main thing I did was avoid poisoning the well. We want to call out
Haskellers' behaviour and propensity to overcomplicate, while at the
same time avoid reinforcing too strongy the perception of by-standers
that using Haskell is a bad idea because (in their view) Haskell
developers overcomplicating evrything is an unavoidable fact of life.