feat: automatically generate media tags for posts
This generates automatically media tags such as those in the "Mapping
Nixpkgs" post.

The description of the post and the image to be shown can be tweaked in
the header of the post if desired. Otherwise, the description is
defaulted to the title, and the image to Tweag's logo.

I changed the last posts accordingly.