Deploy on Netlify from a GitHub action
Previously, we were building the website twice: once in GitHub Actions
to perform linting and check that the website builds, and once in
Netlify to create a deploy preview. We now have a solution to deploy
using the Netlify CLI in GitHub Actions, and annotate the head commit
of each push with the corresponding deployment URL, just like before.

Unlike before, in this PR deployment URL's are immutable. Instead of
one URL per pull-request, there is now one URL per push head commit.
This is due to the fact that PR deployments are not available from the
Netlify CLI. But in any case immutable deployment URL's is the design
adopted by Gatsby Cloud, and it works well there.

The website is now built only once, in an environment that we can
better control than the Netlify build environment. We could even use
the Nix shell in GitHub actions to provision dependencies, just like
locally.