Break out lint into a separate job
In this way, failing the lint does not fail a deploy. We'd rather not
deploy to production code that has lints, but preview deploys should
be fine. Lint can still be made a required check in the GitHub repo
settings, so moving linting into a separate job doesn't mean we're now
allowing lints to ship in production.