Fix javascript errors thrown by invalid contact form attributes
Normally onFocus and onBlur could work like this, but in a React
context the literal strings of JS are not allowed, and runtime errors
result - this is most obvious when, say, right-clicking into a field
in development mode, but happens on the production site too.

Since the existing code doesn't work, removing these two event
handlers doesn't change how the placeholders work, which is simply
according to browser defaults. There's really no need for us to
override them.