Set a minimum width for landing page text
On most screen sizes, "scale your engineering power" is broken on
3 lines. On smaller sizes however, "scale your" ends up on two lines
alone. This doesn't help anything because "engineering" takes up more
width than "scale your". So it's never the right thing for the browser
to lay out the text on 4 lines. To prevent this, we set a `min-width`
that is about the same size as the longest line: the one with
"engineering".

Fixes #631