Add two example YAML profiles
These profiles will be used to generate CV's, which Sales needs on
occasion to close deals. We can leverage Gatsby's data source
facilities to do so. This can also act as the ground truth for bios,
which we need on the team page.

The two examples are Eelco Dolstra and Shao Cheng.