Initial draft of HF backstory (#710)
* Initial draft of HF backstory

* Update 2021-01-02-haskell-foundation-backstory.md

Great initial draft. Thanks @goldfirere! I added some, condensed some, and reordered events slightly according to my memory. Pls reread for content and typos.

I think the links in the epilogue probably need to be updated to reflect Discourse etc. Asking Emily. But we can do that in parallel.

* Small edits on Tim's rewrite

* Update 2021-01-02-haskell-foundation-backstory.md

I addressed all of SPJ's comments more or less. 
1) "mojo" deleted. 
2) Better back story including Frank and Skills Matter
3) Cleaned up the flow around the funding question and added color
4) Cleaned up the Working Group naming 
5) SPJ wanted more pointing to Ethos. Instead of that I went more personal (on behalf of all 3 of us)
6) "gratified" deleted in connection with haskell.org conversations

Also 
1) Added some color from @mboes from our call and changed the headings a little to be less likely to trigger a reader.
2) Probably added some typos! Will reread tomorrow.

It's getting close. Still open to more input if reading it prompts any reactions.

* A few small edits

* Spellcheck; remove redundant bit about Emily

* Update 2021-01-02-haskell-foundation-backstory.md

minor edit

* line break all paragraphs

easier to edit and review.

* edits

* Missing link

* Amend the headings and other assorted edits

I think the pandemic-based titles were too cutesy, and we run the risk
of coming across as insensitive to the harder hit victims of the pandemic.

* lint

* Small edits

* Change post date

Co-authored-by: Tim Sears <tim@timsears.com>
Co-authored-by: Mathieu Boespflug <m@tweag.io>
Co-authored-by: Arnaud Spiwack <arnaud@spiwack.net>
Co-authored-by: guaraqe <mail@guaraqe.gq>