Set margins and padding for h3-5 similarly to h1-2
Smaller headers like h3, h4, and h5 had padding and margin set in
a way which was different than h1 and h2 did.

As a result, we saw odd behavior where, when the viewport was narrow
enough, h3-h5 would not be aligned with the surrounding text: we saw
both a lack of padding (#1131) and an excess of padding (#1151).

By updating the CSS for the blog so h3, h4, and h5 so they have
padding and margin set in the same way that we do for h1 and h2, we
fix weird issues where changing the size of the window causes some
headers to be padded inconsistently.

This change also reduces the bottom margin for h4 and h5 so the break
is a better fit for the reduced size of the font used for those
headers.

A test blog post, to be removed before merging, is included with this
commit so others can verify that this actually fixes the problem.

Please make sure to test with screens/windows of different sizes!